Move travel component to common
