+++ package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@githubprimer/octicons-react": "^8.0.0",
     "bootstrap": "^3.3.7",
-    "d3-scale": "^2.1.0",
     "isomorphic-fetch": "^2.2.1",
     "node-sass-chokidar": "^1.3.3",
     "npm-run-all": "^4.1.3",
